refactor(identity-security-starter): 클래스 이름 및 패키지 위치 수정
