+++ board/board-web-mvc/build.gradle.kts
@@ -10,7 +10,7 @@ dependencies {
     implementation(project(":board:board-application"))
     implementation(project(":board:board-domain"))
 
-    implementation(project(":identity:identity-client-starter"))
+    implementation(project(":identity:identity-security-starter"))
 
     testImplementation(project(":kernel:kernel-test"))
 }
