--- identity/identity-client-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,0 @@
-com.seatliberator.seatliberator.identity.client.IdentityClientAutoConfiguration
-com.seatliberator.seatliberator.identity.client.IdentityClientSecurityAutoConfiguration
