+++ identity/identity-security-starter/docs/NAMESPACE_ROLE_CAPABILITY_EXPANSION_GUIDE.md
