+++ identity/identity-security-starter/src/main/java/com/seatliberator/seatliberator/identity/starter/security/IdentitySecurityStarterOAuth2AutoConfigure.java
@@ -1,9 +1,9 @@
-package com.seatliberator.seatliberator.identity.client;
+package com.seatliberator.seatliberator.identity.starter.security;
 
 import com.seatliberator.seatliberator.bootstrap.security.SecurityStarterAutoConfigure;
 import com.seatliberator.seatliberator.bootstrap.security.customizer.HttpSecurityCustomizer;
-import com.seatliberator.seatliberator.identity.client.jwt.ActorContextBindingFilter;
-import com.seatliberator.seatliberator.identity.client.jwt.ActorContextJwtAuthenticationConverter;
+import com.seatliberator.seatliberator.identity.starter.security.jwt.ActorContextBindingFilter;
+import com.seatliberator.seatliberator.identity.starter.security.jwt.ActorContextJwtAuthenticationConverter;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
@@ -19,7 +19,7 @@ import org.springframework.security.oauth2.server.resource.web.authentication.Be
         havingValue = "true",
         matchIfMissing = true
 )
-public class IdentityClientSecurityAutoConfiguration {
+public class IdentitySecurityStarterOAuth2AutoConfigure {
 
     @Bean("oAuth2ResourceServerCustomizer")
     @ConditionalOnBean({
