+++ identity/identity-security-starter/src/main/java/com/seatliberator/seatliberator/identity/starter/security/jwt/ActorContextAuthenticationToken.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.identity.client.jwt;
+package com.seatliberator.seatliberator.identity.starter.security.jwt;
 
 import com.seatliberator.seatliberator.identity.core.actor.Actor;
 import org.jspecify.annotations.NullMarked;
