+++ settings.gradle.kts
@@ -27,8 +27,8 @@ include("gateway")
 
 include("identity")
 include("identity:identity-core")
-include("identity:identity-client-starter")
 include("identity:identity-application-starter")
+include("identity:identity-security-starter")
 include("identity:identity-server")
 include("identity:identity-server:identity-server-api")
 include("identity:identity-server:identity-server-application")
